+++ src/app/app/day/page.tsx
@@ -24,6 +24,111 @@ import {
   CarouselPrevious,
 } from "@/components/ui/carousel"
 
+const DAILY_ACTIONS = [
+  {
+    name: 'Drink Water',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Shower',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Take meds',
+    area: 'self',
+    categories: ['body'],
+    cadence: '2-daily',
+  },
+  {
+    name: 'Log mood',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Eat breakfast',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Eat lunch',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Eat dinner',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Brush teeth',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Workout',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Use CBD',
+    area: 'self',
+    categories: ['spirituality'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Work',
+    area: 'home',
+    categories: ['work'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Wash dishes',
+    area: 'home',
+    categories: ['clean'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Store dishes',
+    area: 'home',
+    categories: ['clean'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Check trash',
+    area: 'home',
+    categories: ['clean'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Brush floor',
+    area: 'home',
+    categories: ['clean'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Had an orgasm',
+    area: 'self',
+    categories: ['body'],
+    cadence: 'daily',
+  },
+  {
+    name: 'Went out',
+    area: 'social',
+    categories: ['community'],
+    cadence: 'daily',
+  },
+]
+
 
 
 
@@ -53,9 +158,13 @@ export default function Template({ title, content, isomorphicContent }: any) {
       <div className="flex flex-wrap justify-center">
         <Carousel>
           <CarouselContent>
-            <CarouselItem className="flex flex-col">        <label>Friday, Jul 25, 2025</label>
+            <CarouselItem className="flex flex-col text-center">        
+              <small>$280</small>
+              <label>Friday, Jul 25, 2025</label>
+              
         <Button>Close day</Button></CarouselItem>
             <CarouselItem className="flex flex-col"><label>Saturday, Jul 26, 2025</label>
+              <small>$280</small>
         <Button>Close day</Button></CarouselItem>
           </CarouselContent>
           <CarouselPrevious />
